refactor(item-group): move clipboard and edit utils to application layer
- Move clipboardGuards, itemGroupEditUtils, useItemGroupClipboardActions, itemGroupModals to application/
- Update all imports to new absolute paths
- Remove unused imports and fix ESLint errors
- Arquitetura alinhada ao padrão do projeto
