+++ src/modules/diet/item-group/application/clipboardGuards.ts
